Refactor `App` namespace to `Chat`
